+++ database/migrations/users.go
@@ -1,7 +1,7 @@
 package migrations
 
 type Users struct {
-	ID      int32  `gorm:"primaryKey"`
+	ID      int64  `gorm:"primaryKey"`
 	Mobile  string `gorm:"type:varchar(20);unique;not null"`
-	Balance int32  `gorm:"not null"`
+	Balance int64  `gorm:"not null"`
 }
